# api/openapi-bundled.yaml
openapi: 3.1.0
info:
  title: Focus API
  description: Personal task management (projects, tasks, tags, reviews) with Google sign-in and an MCP server.
  version: 1.0.0
  contact:
    name: Cosmic Teacups
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://focus-api.cosmicteacups.com
    description: Production server
security:
  - cookieAuth: []
...
paths:
  /api/auth/google/login:
    get:
      tags:
        - auth
      summary: Start Google sign-in
      description: Redirects the browser to Google. Pass ?redirect=/path to land somewhere specific after sign-in.
      operationId: googleLogin
      parameters:
        - name: redirect
          in: query
          required: false
          description: Frontend path to return to after sign-in
          schema:
            type: string
      responses:
        '302':
          description: Redirect to Google
      security: []
...
  /api/auth/logout:
    post:
      tags:
        - auth
      summary: Log out
      operationId: logout
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - cookieAuth: []
  /api/auth/me:
    get:
      tags:
        - auth
      summary: Current user
      operationId: getMe
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - cookieAuth: []
  /api/user-preferences:
    get:
      tags:
        - auth
      summary: Get user preferences
      operationId: getUserPreferences
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPreferences'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - cookieAuth: []
    patch:
      tags:
        - auth
      summary: Update user preferences
      operationId: updateUserPreferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserPreferencesRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPreferences'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - cookieAuth: []
  /api/api-tokens:
    get:
      tags:
        - auth
      summary: List API tokens
      operationId: listApiTokens
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ApiToken'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - cookieAuth: []
    post:
      tags:
        - auth
      summary: Create an API token
      description: Creates a personal API token for MCP clients. The secret is returned once.
      operationId: createApiToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateApiTokenRequest'
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateApiTokenResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - cookieAuth: []
  /api/api-tokens/{tokenId}:
    delete:
      tags:
        - auth
      summary: Revoke an API token
      operationId: deleteApiToken
      parameters:
        - name: tokenId
          in: path
          required: true
          description: API token ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - cookieAuth: []
  /api/tasks:
    get:
      tags:
        - tasks
      summary: List tasks
      description: Defaults to view=available when no filter is given.
      operationId: listTasks
      parameters:
        - name: view
          in: query
          required: false
          description: Named view. inbox = no project; available = active and not deferred; flagged = urgent; due = has a due date (soonest first); completed = recently completed; all = every active task
          schema:
            type: string
            enum:
              - inbox
              - available
              - flagged
              - due
              - completed
              - all
        - name: project_id
          in: query
          required: false
          description: Only tasks in this project
          schema:
            type: string
            format: uuid
        - name: tag_id
          in: query
          required: false
          description: Only tasks carrying this tag
          schema:
            type: string
            format: uuid
        - name: q
          in: query
          required: false
          description: Case-insensitive title/note search
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - cookieAuth: []
    post:
      tags:
        - tasks
      summary: Create a task
      description: Creates a task. Without project_id it lands in the inbox.
      operationId: createTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - cookieAuth: []
  /api/tasks/{taskId}:
    get:
      tags:
        - tasks
      summary: Get a task
      operationId: getTask
      parameters:
        - name: taskId
          in: path
          required: true
          description: Task ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - cookieAuth: []
    patch:
      tags:
        - tasks
      summary: Update a task
      operationId: updateTask
      parameters:
        - name: taskId
          in: path
          required: true
          description: Task ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - cookieAuth: []
    delete:
      tags:
        - tasks
      summary: Delete a task
      operationId: deleteTask
      parameters:
        - name: taskId
          in: path
          required: true
          description: Task ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - cookieAuth: []
  /api/tasks/{taskId}/complete:
    post:
      tags:
        - tasks
      summary: Complete a task
      description: Marks the task completed. For repeating tasks the next occurrence is created and returned as next_task.
      operationId: completeTask
      parameters:
        - name: taskId
          in: path
          required: true
          description: Task ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompleteTaskResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - cookieAuth: []
  /api/tasks/{taskId}/drop:
    post:
      tags:
        - tasks
      summary: Drop a task
      operationId: dropTask
      parameters:
        - name: taskId
          in: path
          required: true
          description: Task ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - cookieAuth: []
  /api/tasks/{taskId}/reopen:
    post:
      tags:
        - tasks
      summary: Reopen a completed or dropped task
      operationId: reopenTask
      parameters:
        - name: taskId
          in: path
          required: true
          description: Task ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - cookieAuth: []
  /api/tasks/{taskId}/defer:
    post:
      tags:
        - tasks
      summary: Defer a task
      description: Sets defer_until from a natural-language phrase ("1w", "next monday", "in 3 days"), an explicit timestamp, or clears it.
      operationId: deferTask
      parameters:
        - name: taskId
          in: path
          required: true
          description: Task ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeferTaskRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - cookieAuth: []
  /api/tasks/reorder:
    post:
      tags:
        - tasks
      summary: Reorder tasks
      description: Sets sort_order for the given tasks to their position in the list. Used for drag-and-drop in a project or the inbox; the order matters for sequential projects.
      operationId: reorderTasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReorderTasksRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - cookieAuth: []
  /api/tasks/{taskId}/repeat:
    post:
      tags:
        - tasks
      summary: Set a task's repeat schedule
      description: Sets repeat_rule from a natural-language phrase ("first of every month", "every other friday", "every 2 weeks after completion"), an explicit rule, or clears it. Calendar-anchored rules on an undated task also set due_at to the first occurrence.
      operationId: setTaskRepeat
      parameters:
        - name: taskId
          in: path
          required: true
          description: Task ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetRepeatRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - cookieAuth: []
  /api/dates/parse:
    post:
      tags:
        - tasks
      summary: Parse a natural-language date
      description: Resolves a phrase to a timestamp using the rule-based parser, falling back to Claude for anything it cannot handle.
      operationId: parseDate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ParseDateRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParseDateResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - cookieAuth: []
  /api/repeats/parse:
    post:
      tags:
        - tasks
      summary: Parse a natural-language repeat schedule
      description: Resolves a phrase to a repeat rule using the rule-based parser, falling back to Claude, without changing anything.
      operationId: parseRepeat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ParseRepeatRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParseRepeatResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - cookieAuth: []
  /api/counts:
    get:
      tags:
        - tasks
      summary: Sidebar badge counts
      operationId: getCounts
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Counts'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - cookieAuth: []
  /api/projects:
    get:
      tags:
        - projects
      summary: List projects
      description: Flat list ordered by depth then sort_order; nest client-side using parent_id.
      operationId: listProjects
      parameters:
        - name: status
          in: query
          required: false
          description: Filter by status; omit for active + on_hold
          schema:
            type: string
            enum:
              - active
              - on_hold
              - completed
              - dropped
              - all
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Project'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - cookieAuth: []
    post:
      tags:
        - projects
      summary: Create a project
      operationId: createProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - cookieAuth: []
  /api/projects/reorder:
    post:
      tags:
        - projects
      summary: Reorder projects
      description: Sets sort_order for the given projects to their position in the list. Intended for drag-and-drop among siblings in the project tree.
      operationId: reorderProjects
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReorderProjectsRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Project'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - cookieAuth: []
  /api/projects/{projectId}:
    get:
      tags:
        - projects
      summary: Get a project with its tasks and children
      operationId: getProject
      parameters:
        - name: projectId
          in: path
          required: true
          description: Project ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectDetail'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - cookieAuth: []
    patch:
      tags:
        - projects
      summary: Update a project
      operationId: updateProject
      parameters:
        - name: projectId
          in: path
          required: true
          description: Project ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProjectRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - cookieAuth: []
    delete:
      tags:
        - projects
      summary: Delete a project
      description: Deletes the project, its child projects and all of their tasks.
      operationId: deleteProject
      parameters:
        - name: projectId
          in: path
          required: true
          description: Project ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - cookieAuth: []
  /api/projects/{projectId}/review:
    post:
      tags:
        - projects
      summary: Mark a project as reviewed
      description: Stamps last_reviewed_at and schedules the next review review_interval_days later.
      operationId: reviewProject
      parameters:
        - name: projectId
          in: path
          required: true
          description: Project ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - cookieAuth: []
  /api/reviews:
    get:
      tags:
        - projects
      summary: Projects due for review
      operationId: listReviews
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - cookieAuth: []
  /api/tags:
    get:
      tags:
        - tags
      summary: List tags
      operationId: listTags
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tag'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - cookieAuth: []
    post:
      tags:
        - tags
      summary: Create a tag
      operationId: createTag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTagRequest'
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - cookieAuth: []
...
components:
  schemas:
    Error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: Error message
          example: An error occurred
    OkResponse:
      type: object
      required:
        - ok
      properties:
        ok:
          type: boolean
          example: true
...
    MeResponse:
      type: object
      required:
        - user
      properties:
        user:
          $ref: '#/components/schemas/User'
    UserPreferences:
      type: object
      required:
        - timezone
      properties:
        timezone:
          type: string
          example: America/New_York
    UpdateUserPreferencesRequest:
      type: object
      properties:
        timezone:
          type: string
          example: America/New_York
    ApiToken:
      type: object
      description: A personal API token used to authenticate MCP clients. The secret is only returned once, at creation.
      required:
        - id
        - name
        - token_prefix
        - created_at
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: Claude Code on laptop
        token_prefix:
          type: string
          description: First characters of the token, for display
          example: fcs_ab12
        last_used_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
    CreateApiTokenRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          example: Claude Code on laptop
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session_id
      description: Session ID cookie set after Google sign-in